[4/n] Fix CI
I misread the documentation and we were deleting the file
when the context manager exited, which deleted the file
on Windows (but not on other platforms).

The temporary solution is to leave the file in the temp
directory for now. We can work out a better solution when
we have a working script on Windows.